feat: deprecate old api
